UI/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biovision-napari"
version = "0.1.0"
description = "Napari plugin for multimodal bioimage segmentation workflows"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "James DiMartino" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: napari",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "napari>=0.5.0",
    "qtpy",
    "pydantic>=2.0",
    "pyyaml",
    "numpy",
    "dask[array]",
    "zarr>=2.0,<3.0",
    "tifffile",
    "h5py",
    "nd2",
    "readlif",
    "aicsimageio>=4.0",
    "aicspylibczi",   # CZI support for aicsimageio
    "bioio-lif",      # LIF support
    "imagecodecs",    # additional tiff compression codecs
    "anthropic",      # LLM chat (Anthropic)
    "openai",         # LLM chat (OpenAI-compatible)
]

[project.optional-dependencies]
dev = ["pytest", "pytest-qt", "pytest-cov", "napari[pyqt5]"]

[project.entry-points."napari.manifest"]
biovision-napari = "biovision_napari:napari.yaml"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
biovision_napari = ["napari.yaml"]
